refactor: Rename `lib/test/runTest.js`
