+++ lib/test/index.js
@@ -5,7 +5,7 @@ const fse = require('fs-extra');
 const yaml = require('js-yaml');
 const { progress, style, writeText } = require('@serverless/utils/log');
 
-const runTest = require('./runTest');
+const runTest = require('./run-test');
 
 module.exports.test = async (ctx) => {
   if (!ctx.sls.enterpriseEnabled) {
