+++ lib/test/index.test.js
@@ -4,7 +4,7 @@ const { expect } = require('chai');
 const runServerless = require('../../test/run-serverless');
 
 const modulesCacheStub = {
-  '@serverless/dashboard-plugin/lib/test/runTest': async () => {},
+  '@serverless/dashboard-plugin/lib/test/run-test': async () => {},
   [require.resolve('@serverless/utils/config')]: {
     ...require('@serverless/utils/config'),
     getLoggedInUser: () => ({
@@ -85,7 +85,7 @@ describe('test', () => {
           command: 'test',
           modulesCacheStub: {
             ...modulesCacheStub,
-            '@serverless/dashboard-plugin/lib/test/runTest': async (testSpec) => {
+            '@serverless/dashboard-plugin/lib/test/run-test': async (testSpec) => {
               if (testSpec.name === 'function') {
                 throw Object.assign(new Error('Fail'), {
                   resp: { headers: {} },
