+++ lib/test/run-test.test.js
@@ -31,7 +31,7 @@ describe('runTest', () => {
         headers: { _headers: {} },
       };
     });
-    runTest = proxyquire('./runTest', {
+    runTest = proxyquire('./run-test', {
       'node-fetch': fetch,
     });
   });
